- replace all \0 characters in Dehacked patches with spaces.
This is needed to make some old and broken patches in some WolfenDoom mods load.

+++ src/c_dispatch.cpp
@@ -694,7 +694,7 @@ void C_DoCommand (const char *cmd, int keynum)
 			}
 			else
 			{
-				Printf ("Unknown command \"%.*s\"\n", len, beg);
+				Printf ("Unknown command \"%.*s\"\n", (int)len, beg);
 			}
 		}
 	}
